test(worker): consumers
Put mocks into mocks folder
